power: oneplus: Optimize coding
* Remove initialize statics.
* Remove else after return.
* Remove Duplicate static void switch_mode_to_normal
  & extern from oem_external_fg.h.
* Remove unnecessary 'out of memory' message.
* Improves dashchg_firmware_data Matrix.
* Use octal permissions.
* Add irq_set_status_flags(di->irq, IRQ_DISABLE_UNLAZY);.
* Convert msleep to usleep_range.
* Replace -EIO to 0, We're on 3.18

And other Optimize coding.

+++ drivers/power/op_dash_adapter.c
@@ -540,7 +540,7 @@ update_err:
 	return false;
 }
 
-bool dash_adapter_update_is_tx_gpio(unsigned long gpio_num)
+bool dash_adapter_update_is_tx_gpio(unsigned int gpio_num)
 {
 	if (!the_chip)
 		return false;
@@ -550,7 +550,7 @@ bool dash_adapter_update_is_tx_gpio(unsigned long gpio_num)
 		return false;
 }
 
-bool dash_adapter_update_is_rx_gpio(unsigned long gpio_num)
+bool dash_adapter_update_is_rx_gpio(unsigned int gpio_num)
 {
 	if (!the_chip)
 		return false;
